% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_i_gexp_sper_ont_trans}
\alias{cell_i_gexp_sper_ont_trans}
\title{Introns matrix of mouse spermary cells using HISOFA-seq with gene clustering}
\format{
A matrix with 17721 rows and 3000 columns
\describe{
  ...
}
}
\source{
\url{https://github.com/Dekayzc/Regionvelocity}
}
\usage{
cell_i_gexp_sper_ont_trans
}
\description{
Introns matrix of mouse spermary cells containing gene and cell name
rows are genes and columns are cells.
}
\keyword{datasets}
