% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_ls_gexp_sper_pb}
\alias{cell_ls_gexp_sper_pb}
\title{Length of spliced counts matrix of mouse spermary cells using HIT-scISOseq}
\format{
A matrix with 18255 rows and 3000 columns
\describe{
  ...
}
}
\source{
\url{https://github.com/Dekayzc/Regionvelocity}
}
\usage{
cell_ls_gexp_sper_pb
}
\description{
Length matrix of mouse spermary cells containing gene and cell name
rows are genes and columns are cells, length is total length.
}
\keyword{datasets}
