% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_u_gexp_eyes}
\alias{cell_u_gexp_eyes}
\title{Unspliced counts matrix of monkey limbal cells}
\format{
A matrix with 17408 rows and 1417 columns
\describe{
  ...
}
}
\source{
\url{https://github.com/Dekayzc/Regionvelocity}
}
\usage{
cell_u_gexp_eyes
}
\description{
Unspliced counts matrix of monkey limbal cells containing gene and cell name
rows are genes and columns are cells.
}
\keyword{datasets}
