% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_u_gexp_sper_ont_gene}
\alias{cell_u_gexp_sper_ont_gene}
\title{Unspliced counts matrix of mouse spermary cells using HISOFA-seq without gene clustering}
\format{
A matrix with 19221 rows and 3000 columns
\describe{
  ...
}
}
\source{
\url{https://github.com/Dekayzc/Regionvelocity}
}
\usage{
cell_u_gexp_sper_ont_gene
}
\description{
Unspliced counts matrix of mouse spermary cells containing gene and cell name
rows are genes and columns are cells.
}
\keyword{datasets}
